refactor: clearing all users from database exclues admins

+++ project-backend/src/data/db/dbUser.ts
@@ -65,5 +65,5 @@ export const getAllUsers = async () => {
 };
 
 export const clearUsers = async () => {
-  await User.deleteMany({});
+  await User.deleteMany({ isAdmin: { $ne: true }});
 };
